Taking a Break
I've been working on this little project for a bit but I think I'm juggling too many hats at once and thats killing my motivation and enjoyment for the project.
I'm going to of course document where I'm at with this right now but I'm going to start fresh and try to do things in little sprints with greater feedback for me.